feat(phase-5): LLM 기반 스크립트 분석 도구 구현
Phase 5 완료 - 다양한 형식의 텍스트(소설, 시나리오, 묘사 등)를
자동으로 분석하여 캐릭터, 장면, 대사, 관계 정보를 추출하는
LLM 기반 스크립트 분석 도구를 구현했습니다.

주요 변경사항:

Backend (api-server):
- Script 엔티티 추가 (업로드된 스크립트 저장)
- ScriptRepository, ScriptService, ScriptController 구현
- POST /scripts/upload-and-analyze 엔드포인트 추가
- 스크립트 업로드 및 분석 결과 저장 기능
- gradle.properties 수정 (Java 21 명시적 설정)

Frontend:
- /script-analyzer 페이지 구현
- 스크립트 입력 및 분석 결과 시각화 UI
- 형식 선택 (소설, 시나리오, 묘사, 대화)
- LLM 프로바이더 선택 (OpenAI, Claude, Gemini)
- 홈페이지 그리드 레이아웃 4컬럼으로 변경

LLM Server:
- POST /gen/analyze-script 엔드포인트 추가
- 캐릭터 자동 추출 (이름, 성격, 말투, 대사 예시)
- 장면 정보 추출 (위치, 분위기, 참여자)
- 대사 추출 및 화자 매칭
- 캐릭터 간 관계 분석 (관계 유형, 친밀도)
- JSON 파싱 개선 (마크다운 코드 블록 처리)

Bug Fixes:
- RelationshipService.java 타입 에러 수정 (Integer → Double)

Documentation:
- NEXT_TASKS.md 업데이트 (Phase 5 완료 상태 반영)
- README.md 업데이트 (Phase 5 구현 내용 추가)

+++ api-server/src/main/java/com/jwyoo/api/service/RelationshipService.java
@@ -105,7 +105,7 @@ public class RelationshipService {
         Relationship relationship = getRelationshipById(id);
 
         String oldType = relationship.getRelationType();
-        Integer oldCloseness = relationship.getCloseness();
+        Double oldCloseness = relationship.getCloseness();
 
         relationship.setRelationType(updatedRelationship.getRelationType());
         relationship.setCloseness(updatedRelationship.getCloseness());
